The send and receive pages assumed merit was returned from rateservice.fromFiat
This caused an obviously bad bug since RateService.fromFiat returns micros.
I change toFiat and fromFiat to include units so that there won't be confusion
in the future. The send and receive pages were fixed.

Also the qr code address uses "micros" in url instead of amount so that
it's more clear what unit is there.