Fix styling of wallet amount in details page and fix for icon in overview.
The wallet amount hid under the transaction list when you shrunk the width
of the window. This fix makes sure it is always shown at the top.

Also the icon shown for a wallet in the main view was way too small and kind
of grey. This now fixes the size to be larger and obviously nicer.